Add skeleton for collection transformation
Implement a base framework for transforming collection data, providing
base64 en/decoding and lowercase a transformations as a start. Transformations
are not cached in ctx.collections. Transformations are defined as
a key/value pair in the rule opts table.

Also fixes a small bug in _equals that could lead to a 500