Implement rule offset pre-calculation and handling
Move rule processing from a naive iteration loop to an
offset-based flow as suggested by @splitice in #39. Note:

* Offsets are calculated by FW.init() in `init_by_lua`
* Rule tables are now global, and no longer wrapped in an accessor
* No more skipend/chainchild/chainend flag nonsense

Rule processing now moves forward by taking the offset of the rule
(offsets for match vs. nomatch will differ for CHAIN and SKIP rules)
and moving a pointer forward. Offset values are stored in the rule
table itself as top-level keys of the rule, and are transparent to
both the user and the rule definition.