Break utility functions and lookup tables to individual modules
Create a number of separate libraries to hold helper functions
and lookup tables. These libraries include:

* lib/log: Debug and fatal fail logging
* lib/lookup: Lookup tables
* lib/operators: Rule operators (not all operators live here)
* lib/rule_calc: Existed from a previous commit, changes here are whitespace
* lib/storage: Shared dict persistent storage
* lib/util: Misc functions, mostly table manipulation

Because most functions contained debug log calls, they required a
reference to the 'fw' object. Moving _log(...) to a separate
module means that the signatures of functions don't need to change,
but the previous name of the 'self' reference has been changed to
'fw' to more accurately represent what is being passed in.

Fixes issue #70.