Shuffle collections table and related functions to accommodate phases
Since we will be assigning values to the collections table in
multiple phases, move it to a lookup table. Relatedly,
move several functions related to parsing to a separate module.

Also, we're now storing the collections table in ctx (which is
then saved to the per-request ngx.ctx table). This requires
renaming the transform memoization tables for sanity.