Implement RESPONSE_BODY collection
Use the streaming response body processing made available via
ngx.arg[1] and ngx.arg[2] to build the RESPONSE_BODY collection.
Because this can be very memory intensive, create an option
to define a size threshold, beyond which the response will not be
added as a collection. The size is defined by the Content-Length
header; if this header doesn't exist, fail forward by ignoring
the response body as if the size were beyond the threshold.

Also create an option to define approprite MIME types for response
body analysis (defaulting to text/plain and text/html). As with
the maximum size, this is defined by the Content-Type header. The
same logic applies if the header does not exist.