Implement user-defined rulesets
Add an option `add_ruleset` to allow users to specify additional
Lua modules to be processed as rulesets. Rulesets are processed
in numeric order, so there is a restriction that values must be
number types.

Thanks to @mickygough for the original patchset (which had to
be closed due to merge conflict) and @jfchevrette for +1'ing
this feature.