Implement rule definitions as JSON
Big change that encompasses several scopes, but doing it in one
go makes things a lot cleaner. Rule files are now defined as .json
files, not Lua modules. This means rulesets are no longer require'd,
but rather read from disk via Lua's native I/O library and parsed.
For default rulesets and default_option calls, this is done in the
init_by_lua phase, where blocking I/O is safe.

Because we're keeping the default_option/set_option logic in place,
this means that scopes that add a ruleset outside the init phase
are using blocking I/O in the event loop. Yes, this is a no-no,
but for simplicity's sake we can allow this and document a big fat
warning about this behavior (and the reality is the impact from this,
barring some sort of underlying filesystem failure, is almost nil).

Additionally, new negated operator syntax is now in place, and we've
re-added function-type collections (added in this commit to get tests
working).