Standardize operator returns (1/2)
Standardize the return signature of operator functions to two elements:
'match', a boolean indicating if the collection matched the pattern, and
'value', the portion of the collection which matched against the pattern.

This will allow us to be more flexible with acting on match elements,
as noted in issue #112.

This also allows us to ensure that log entries will always contain
a value associated with the match in the 'match' key (as opposed to
a potential boolean).

This is a two-part commit as the EXISTS and CONTAINS operators need
to be refactored and wrapped in separate functions.