Refactor collection parse logic
Update the parse collection logic to use a single key/value pair
for defining how to parse collection. The use of an iterator to
peek into the first (what should be the only) element of the
sub-table allows us to keep the use of a lookup table, while
simplifying the definition syntax.

Fixes issue #123.