Refactor variable storage for ModSecurity compatibility
Big commit to rework how transaction and persistent variables
are stored, retrieved, and persisted via Nginx shms.

* To provide compability with ModSecurity definitions, variables now
live within a collection such as TX, IP, etc. Collections are initialized
in the same manner as ModSecurity (via the initcol option on rule match).
The TX collection is automatically initialized to itself or an empty table
at the beginning of fw:exec().
* Variables are now stored in a sub-ctx table, and are accessed via their
respective collection names as a var type. For the TX collection, this is
a simple 1-1 reference. Collections whose names are initialized based on
a distinct value (such as IP) are assigned to the collections lookup table
as part of the initcol process.
* Variable collections can be parsed by _parse_collection just as any
other collection, however, these results will never be cached.
* The SETVAR and SETTX actions have been removed; variable setting
is now handled via a rule opt (this is processed when the rule matches).
This brings this functionality to parity with ModSecurity.
* Persistent collections are still stored to ngx shared memory structures,
with the name of the collection as the key, and a serialized representation
of the collection as the value. Collections are not re-persisted if no
values in the collection were re-set.
* Expiration is handled by examining each key in the collection during
initialization and comparing an associated timestamp. Because we store
the entire collection as a serialized table, we cannot rely on ngx.shared's
native expiration mechanism.
* Dynamic parse syntax received a small tweak to meet ModSecurity parity.
* The collection holding the client IP address has been renamed from
IP to REMOTE_ADDR to accomodate the IP variable collection.

Fixes issue #113.