Persist variable storage when keys have expired
Storage collections should be persisted when an expired key is
encountered during initialization. Although not removing the key
would not have resulted in a logic error (because the key was removed
from the memory buffer before being assigned to the collections
lookup), doing so could result in stale keys laying around for a
potentially unbound amount of time.