Auto-populate TX.# collection elements
Populate values in the TX collection based on the results of
value returned from the operators lookup. When REGEX is used,
populate members of the collection based on the capture groups
defined in the pattern.

This requires an adjustment to how we handle 'specific' values when
parsing ModSecurity rules to handle cases such as 'TX:0'.