Intelligently translate auto-expanding and pattern-altering operators
Add translations for for beginsWith, containsWord, and endsWith. These
operators are similar enough to existing functions that we don't
need to write separate wrappers for them; instead, we can mangle
the pattern as appropriate in the translation via the lookup table.

Additionally, we should set the parsepattern flag for certain operators
to meet functionality with their ModSecurity equivalent.

This is part of the list of commits for issue #135.