Refactor offset calcs for chains to handle skipAfter action
The nature of skipAfter (translated to skip_after) requires that we
peek ahead to figure out the offsets for the appropriate skips.
Additionally, implementing skip and skipAfter translations with
chain members requires us to shuffle around rule options in the
location of the chain (such elements are required to live in the
chain started in ModSecurity, but in the last chain element in
FreeWAF).