Implement RBL DNS lookup
Use lua-resty-dns lib to wrap RBL DNS query lookups. Because the
underlying DNS wrapper wants its own defined set of nameservers
to query, we need to provide those via the *_option interface.

Also worth noting that the unit test for the rbl_lookup operator
are very fragile because we're hardcoding a third party rbldns
server, largely out of laziness. The best way to handle this would
be to mock out a local rbldns daemon (lua-resty-dns already does this).

Fixes issue #73.