Implement redis backend for persistent storage
We're using a slightly different approach to writing data to the
backend. Because individual hash keys cannot be expired, we still
need an O(n) loop on initialization to lazily remove expired keys.
Such keys, and keys deleted with `delete_var`, are pushed to a table
to be removed via hdel() when persist is called. Similarly, setting
variables is done by pushing the k/v pair to a hash and calling
hmset().

Additionally, we need to do a type check on intialization because all
values returned from redis are treated as strings.