Add periodic flush call to synopsis example
This has come up a few times (see #183 and #141); users are confused
when setting up test installations and not immediately seeing event
log entries in file or socket. Let's add this to the synopsis to
hopefully cut down on some confusion.

It may also be worth investigating an unbuffered event log option.