Force upper case naming of variables in translation
So... this is annoying. It seems that ModSecurity is case-insensitive
with respect to handling variable names and collections in nondisruptive
actions, but not necessarily as part of collection parsing. For example,
examine the DOS_BLOCK variable found in CRS v3 rc1, where it is
used in both all-upper and all-lower case instances (and yet other
variables are used in collection parsing as lowercased strings).

To simplify this, we're forcing a single case across variable names
in variable-related actions and in macro translation (using the
storage key in the collection lookup table). This is not particularly
elegant, but it's relatively simply; trying to move this logic into
the rule engine's collection lookup, dynamic value parsing, and other
places would be a lot uglier.