Bugfix in collections length handling
Move length modifier handling out of the cache miss logic branch,
and remove the length modifier from collection cache key construction.
This fixes a bug wherein the length modifier was not applied to
storage variable elements (because we always forced a lookup).

Additionally, collections with a length modifier now always have
a truthy value, even if that value is 0, to more accurately reflect
the original meaning of the operator. Thus, length-modified collections
will always be parsed and examined by the rule's operator; this also
more accurately reflects the intention behind some CRS rules.