Define default CRS configuration values
Relying on the SecRules mechanism would be a waste of cycles to
regenerate the same TX values on every request. Instead, we pre-
populate the TX collection with CRS configuration values.

Not all REQUEST-901-INITIALIZATION values are present; those ommited
are related to unsupported operators/actions, or are complex enough
that the user should define them manually (e.g. real_ip).