Do not error when capture won't take affect
See https://github.com/SpiderLabs/owasp-modsecurity-crs/pull/716

Documentation on capture is a bit thin, and behaviors change between
ModSecurity versions. For now, we don't need to warn when we see this
where we won't take action on it, as there are unhandled (but not
behavior-breaking) cases where capture should be set for non-rx
definitions.